after class

+++ rt-spark-engine/src/main/scala/model/Reading.scala
@@ -5,7 +5,7 @@ import java.util.Date
 import java.text.SimpleDateFormat
 
 
-case class Reading (val userId : String, val trajectoryId: String, val lat: Float, val lon : Float, val alt: Int, val date : String, val time : String) extends java.io.Serializable {
+case class Reading (val userId : String, val trajectoryId: String, val lat: Float, val lon : Float, val alt: Float, val date : String, val time : String) extends java.io.Serializable {
 
 	val ts = new SimpleDateFormat("yyyy-MM-dd,HH:mm:ss").parse(date + "," + time).getTime
